// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  STUDENT: 'STUDENT'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const RaffleStatus = {
  DRAFT: 'DRAFT',
  ACTIVE: 'ACTIVE',
  FINISHED: 'FINISHED',
  DRAWN: 'DRAWN',
  CANCELLED: 'CANCELLED'
} as const

export type RaffleStatus = (typeof RaffleStatus)[keyof typeof RaffleStatus]


export const TicketStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  ASSIGNED: 'ASSIGNED'
} as const

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus]


export const PaymentStatus = {
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]